Various parity changes and additions
Make the look of the game more accurate by editing the original textures to fit the colors of the original maze better. The floor and ceiling are now separate objects that are added from the code. I made all of the objects that are loaded from the code use no blurry filter making them look bad. I yet to figure out how to do this on the walls which are the only part of the model now.