W.I.P Collision! plus some tweaks and improvements
Collision is checked every frame through a function in /lib/surface,js
the collision it is self is defined in a level's collision.js file in their respective root directories, they are in the form of cubes that are added around the maze that stop the player if they collide with it.

Besides that, the player movement is now determined through velocity instead of adding an amount to one of the position axes.

There is a minor change with the model where now the wall textures are finally are on point and have no filter.