Wall sliding, code cleanup and optimization
I was able to implement wall sliding meaning now you can walk towards walls and instead of stopping completely you are still moving, just against it. I also added another collision box but with this came a bug where the start logo at the start of the maze clips in and out of view depending on how you look at it due to the collision boxes using an invisible material.

I optimized and refactored some of the backend, it is virtually the same on the front-end but the code is more organized.

I removed the player script and instead used a system more similar to luphoria's Namco Museum JS, a player action state system.